feat(aapd-1504): refactor viewdata from dynamic list controller, add initial has v2 task list
