+++ packages/forms-web-app/src/dynamic-forms/controller.js
@@ -1,12 +1,9 @@
 // common controllers for dynamic forms
-
-const { getLPAUserFromSession } = require('../services/lpa-user.service');
 const { SECTION_STATUS } = require('./section');
 const { getJourney } = require('./journey-factory');
 const logger = require('../lib/logger');
 const ListAddMoreQuestion = require('./dynamic-components/list-add-more/question');
 const questionUtils = require('./dynamic-components/utils/question-utils');
-const { LPA_USER_ROLE } = require('@pins/common/src/constants');
 
 /**
  * @typedef {import('@pins/common/src/dynamic-forms/journey-types').JourneyType} JourneyType
@@ -75,21 +72,15 @@ function buildSectionRowViewModel(key, value, action) {
 }
 
 /**
- * @type {import('express').Handler}
+ * @param {import('express').Request} req
+ * @param {import('express').Response} res
+ * @param {string} pageCaption
+ * @param {object} viewData
  */
-exports.list = async (req, res) => {
+exports.list = async (req, res, pageCaption, viewData) => {
 	//render check your answers view
 	const journeyResponse = res.locals.journeyResponse;
 	const journey = getJourney(journeyResponse);
-	const referenceId = res.locals.journeyResponse.referenceId;
-
-	const user = getLPAUserFromSession(req);
-	const encodedReferenceId = encodeURIComponent(referenceId);
-	const appeal = await req.appealsApiClient.getUsersAppealCase({
-		caseReference: encodedReferenceId,
-		userId: user.id,
-		role: LPA_USER_ROLE
-	});
 
 	const summaryListData = {
 		sections: [],
@@ -132,11 +123,11 @@ exports.list = async (req, res) => {
 	}
 
 	return res.render(journey.listingPageViewPath, {
-		appeal,
+		...viewData,
+		pageCaption,
 		summaryListData,
 		journeyComplete: journey.isComplete(),
 		layoutTemplate: journey.journeyTemplate,
-		pageCaption: `Appeal ${appeal.caseReference}`,
 		journeyTitle: journey.journeyTitle
 	});
 };
