+++ packages/forms-web-app/src/routes/appellant-submission/submission-form.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const {
+	list,
 	question,
 	save,
 	remove,
@@ -14,6 +15,15 @@ const dynamicReqFilesToReqBodyFiles = require('../../dynamic-forms/middleware/dy
 const getJourneyResponse = require('../../dynamic-forms/middleware/get-journey-response-for-appellant');
 
 const router = express.Router();
+
+/**
+ * @type {import('express').Handler}
+ */
+const householderTaskList = async (req, res) => {
+	return list(req, res, 'Householder Appeal', {});
+};
+
+router.get('/', getJourneyResponse, householderTaskList);
 
 router.get('/submit/declaration', getJourneyResponse, appellantSubmissionDeclaration);
